Add files via upload
How to use
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Run Handler after following the instructions in README(which are really easy to implement). Then, be in front of your camera, write your name when the program asks it(in the prompt). After inputting one's name, the program will automatically capture 400 images in 5-7 seconds. After that the images go under some procedures such as face detection, resizing and hog generation. When you are done taking people into the database, just type "done" to the prompt. A classifier which is a multi-class support vector machine will be generated and trained. The trained version can be found as a function in the testTrainedClassifier class.